refactor(web): extract design-system-project pure helpers into the slice
Moves the ~65 module-scope pure functions that DesignSystemProjectPanel
relies on (section building/categorization/status derivation, card-manifest
parsing, kit color hex resolution, generation-step/progress computation,
todo/file-op section-activity matching, and the markdown-document template
helpers) into features/file-workspace/rules.ts + types.ts, along with the
DesignSystemReview*/DesignSystemProjectSection*/DesignSystemGenerationStep*
types and TranslateFn. FileWorkspace.tsx imports them back through the slice
barrel; DesignSystemProjectPanel/DesignSystemProjectLoading/
DesignSystemInlinePreview stay in place for a follow-up pass (still stateful,
still transport-touching).

Behavior-preserving move only; no logic changes. 5,461 -> 4,349 lines.
90 new pure-function tests, all passing against the moved code unmodified.
