+++ apps/web/src/components/FileWorkspace.tsx
@@ -40,21 +40,18 @@ import {
   writeProjectBase64File,
   writeProjectTextFile,
 } from '../providers/registry';
-import type { Dict } from '../i18n/types';
 import { setPendingDesignSystemCreateEntry } from '../analytics/ds-create-entry';
 import { navigate } from '../router';
 import { downloadDesignSystemArchive, downloadProjectArchive } from '../runtime/exports';
 import { finalizeBrandProject } from '../runtime/brands';
-import { deriveFileOps, type FileOpEntry } from '../runtime/file-ops';
-import { parseDesignMd } from '../runtime/design-md-parse';
+import { deriveFileOps } from '../runtime/file-ops';
 import {
   deleteBrandImage,
   deleteBrandLogo,
   readDesignMd,
-  replaceDesignMdColorAtIndex,
   updateBrandColor,
 } from '../runtime/kit-edit';
-import { latestTodosFromEvents, type TodoItem } from '../runtime/todos';
+import { latestTodosFromEvents } from '../runtime/todos';
 import { deliverableSlideNavForActiveFile, isSlideNavDeliverableNow } from '../runtime/slide-nav';
 import { buildSrcdoc } from '../runtime/srcdoc';
 import { useDesignKit, hostnameOf, type KitColor } from '../runtime/design-kit';
@@ -131,29 +128,111 @@ import {
   browserTabIndex,
   browserTabsFromState,
   BROWSER_TAB_PREFIX,
+  buildDesignSystemReviewSections,
+  colorHexFromBrandJson,
+  colorHexFromDesignMd,
   DESIGN_FILES_TAB,
+  DESIGN_SYSTEM_CARD_MANIFEST_OPTIONAL_STRING_FIELDS,
+  DESIGN_SYSTEM_GUIDANCE_FILES,
+  DESIGN_SYSTEM_IMAGE_OR_FONT_EXTENSIONS,
   DESIGN_SYSTEM_TAB,
+  designMdBodyWithColor,
+  designSystemBasename,
+  designSystemFallbackReviewSections,
+  designSystemFileOpBelongsToSection,
+  designSystemGenerationProgress,
+  designSystemGenerationReviewHasStarted,
+  designSystemGuidanceSort,
+  designSystemHasSourceContext,
+  designSystemInitialGenerationSteps,
+  designSystemManifestCardError,
+  designSystemPathMatchesSection,
+  designSystemRelatedFilesForCategory,
+  designSystemReviewAgentTaskLabel,
+  designSystemReviewArtifactSort,
+  designSystemReviewCategoryRank,
+  designSystemReviewGroups,
+  designSystemReviewNeedsAttention,
+  designSystemReviewPreviewDisplay,
+  designSystemReviewSubtitle,
+  designSystemReviewTimeLabel,
+  designSystemReviewTitleFromPath,
+  designSystemSectionActivity,
+  designSystemSectionActivityLabel,
+  designSystemSectionChangedAfterReview,
+  designSystemSectionEditableFile,
+  designSystemSectionPhaseLabel,
+  designSystemSectionPreviewFile,
+  designSystemSectionRunningNotice,
+  designSystemSectionStatus,
+  designSystemSectionStatusClass,
+  designSystemSectionStatusLabel,
+  designSystemSectionTodo,
+  designSystemSectionVisibleDuringGeneration,
+  designSystemTodoActivityPhase,
+  designSystemTodoBelongsToSection,
+  designSystemTodoRank,
+  designSystemTodoSearchText,
+  documentTemplateScenarioKey,
+  formatWorkspaceSnapshotElapsed,
+  inferDesignSystemReviewCategory,
+  initialDesignKitColorHex,
+  initialMarkdownDocument,
   isBrowserTabId,
+  isDesignSystemAssetFile,
+  isDesignSystemEvidenceFile,
+  isDesignSystemGuidanceFile,
+  isDesignSystemPreviewFile,
+  isDesignSystemRawAssetFile,
+  isDesignSystemReviewArtifactFile,
+  isDesignSystemReviewableAssetArtifact,
+  isDesignSystemTokenFile,
+  isDesignSystemUiKitEntryPage,
+  isDesignSystemUiKitFile,
   isLiveArtifactImplementationPath,
   isSketchName,
+  joinProjectFilePath,
   lastWorkspaceTabId,
   maxBrowserTabSequence,
+  nextMarkdownDocumentPath,
+  normalizeDesignKitHex,
+  normalizeDesignSystemPath,
+  normalizeProjectFilePath,
+  optionalDesignSystemManifestString,
   orderWorkspaceTabs,
   parentDirForProjectFile,
+  parseDesignSystemCardManifest,
+  parseDesignSystemCardManifestEntry,
+  preferPreviewArtifactsOverRawAssets,
   reanchorBrowserTabsToCurrentOrder,
   sameFileName,
   scrollWorkspaceTabsWithWheel,
+  slugForTestId,
   Tab,
   tabDropEdgeFromEvent,
+  truncateDesignSystemActivityText,
   type BrowserWorkspaceTab,
+  type DesignSystemCardManifestEntry,
+  type DesignSystemCardManifestMap,
+  type DesignSystemGenerationStep,
+  type DesignSystemProjectSection,
+  type DesignSystemProjectSectionReview,
+  type DesignSystemReviewAgentTask,
+  type DesignSystemReviewCategory,
+  type DesignSystemReviewDecision,
+  type DesignSystemReviewDetails,
+  type DesignSystemReviewEntry,
+  type DesignSystemReviewPreviewDisplay,
+  type DesignSystemSectionActivity,
+  type DesignSystemSectionActivityPhase,
+  type DesignSystemSectionStatus,
   type TabDropEdge,
+  type TranslateFn,
   type WorkspaceOrderedTab,
 } from '../features/file-workspace';
 // Re-exported so `./FileWorkspace` stays a stable import path for the
 // existing test suite now that this helper lives in the slice.
 export { scrollWorkspaceTabsWithWheel };
-
-type TranslateFn = (key: keyof Dict, vars?: Record<string, string | number>) => string;
 
 interface Props {
   projectId: string;
@@ -396,32 +475,6 @@ export interface BrowserAttentionRequest {
   action: 'download-page';
   nonce: number;
 }
-type DesignSystemReviewDecision =
-  NonNullable<ProjectMetadata['designSystemReview']>[string]['decision'];
-type DesignSystemReviewEntry = NonNullable<ProjectMetadata['designSystemReview']>[string];
-type DesignSystemReviewAgentTask = NonNullable<DesignSystemReviewEntry['agentTask']>;
-interface DesignSystemReviewDetails {
-  feedback?: string;
-  files?: string[];
-  agentTask?: DesignSystemReviewAgentTask;
-}
-type DesignSystemSectionStatus =
-  | 'missing'
-  | 'planned'
-  | 'running'
-  | 'needs-review'
-  | 'approved'
-  | 'needs-work'
-  | 'updated';
-type DesignSystemReviewCategory = 'Type' | 'Colors' | 'Spacing' | 'Components' | 'Brand';
-interface DesignSystemProjectSection {
-  title: string;
-  subtitle: string;
-  files: string[];
-  category: DesignSystemReviewCategory;
-  requiredFile?: string;
-}
-
 interface SaveSketchOptions {
   activate?: boolean;
   refreshFiles?: boolean;
@@ -454,23 +507,6 @@ function consumeFileWorkspaceTabShortcut(event: KeyboardEvent) {
   event.stopPropagation();
 }
 
-type DesignSystemSectionActivityPhase =
-  | 'idle'
-  | 'planned'
-  | 'reading'
-  | 'writing'
-  | 'updated'
-  | 'error';
-interface DesignSystemSectionActivity {
-  running: boolean;
-  mutated: boolean;
-  errored: boolean;
-  phase: DesignSystemSectionActivityPhase;
-  touchedFiles: string[];
-  todoText?: string;
-  todoStatus?: TodoItem['status'];
-}
-
 function formatBrowserTabUrl(url: string): string {
   if (!url) return '';
   try {
@@ -499,41 +535,6 @@ function createDefaultDesignFilesNavState(): DesignFilesNavState {
   };
 }
 
-interface DesignSystemProjectSectionReview {
-  section: DesignSystemProjectSection;
-  previewFile: ProjectFile | null;
-  previewDisplay: DesignSystemReviewPreviewDisplay;
-  reviewEntry: DesignSystemReviewEntry | undefined;
-  sectionActivity: DesignSystemSectionActivity;
-  changedAfterFeedback: boolean;
-  sectionStatus: DesignSystemSectionStatus;
-  sectionStatusLabel: string;
-  reviewTimeLabel: string | null;
-}
-type DesignSystemReviewPreviewDisplay = 'specimen' | 'ui-kit' | 'asset';
-interface DesignSystemCardManifestEntry {
-  path: string;
-  group?: string;
-  name?: string;
-  subtitle?: string;
-  viewport?: string;
-}
-type DesignSystemCardManifestMap = Map<string, DesignSystemCardManifestEntry>;
-const DESIGN_SYSTEM_CARD_MANIFEST_OPTIONAL_STRING_FIELDS = ['group', 'name', 'subtitle', 'viewport'] as const;
-type DesignSystemGenerationStepStatus = 'pending' | 'running' | 'succeeded';
-interface DesignSystemGenerationStep {
-  id: string;
-  title: string;
-  detail: string;
-  status: DesignSystemGenerationStepStatus;
-}
-const DESIGN_SYSTEM_GUIDANCE_FILES = new Set([
-  'design.md',
-  'readme.md',
-  'readme-print.md',
-  'skill.md',
-]);
-const DESIGN_SYSTEM_IMAGE_OR_FONT_EXTENSIONS = /\.(svg|png|jpe?g|gif|webp|avif|ico|otf|ttf|woff2?)$/i;
 
 type WorkspaceToastTone = 'default' | 'success' | 'error' | 'loading';
 
@@ -4095,1122 +4096,6 @@ function DesignSystemProjectLoading({
       </div>
     </div>
   );
-}
-
-function normalizeDesignKitHex(value: string): string | null {
-  const trimmed = value.trim();
-  const withHash = trimmed.startsWith('#') ? trimmed : `#${trimmed}`;
-  if (/^#[0-9a-fA-F]{6}$/.test(withHash)) return withHash.toUpperCase();
-  if (/^#[0-9a-fA-F]{3}$/.test(withHash)) {
-    return `#${withHash[1]}${withHash[1]}${withHash[2]}${withHash[2]}${withHash[3]}${withHash[3]}`.toUpperCase();
-  }
-  return null;
-}
-
-function initialDesignKitColorHex(
-  index: number,
-  sources: {
-    brandJson: string | null;
-    designMdBody: string | null;
-    swatches: string[] | undefined;
-    currentColors: KitColor[];
-  },
-): string | null {
-  const brandColor = colorHexFromBrandJson(sources.brandJson, index);
-  if (brandColor) return brandColor;
-  const designMdColor = colorHexFromDesignMd(sources.designMdBody ?? '', index);
-  if (designMdColor) return designMdColor;
-  return normalizeDesignKitHex(sources.swatches?.[index] ?? sources.currentColors[index]?.hex ?? '');
-}
-
-function colorHexFromBrandJson(raw: string | null, index: number): string | null {
-  if (!raw) return null;
-  try {
-    const parsed = JSON.parse(raw) as { colors?: Array<{ hex?: unknown }> };
-    const hex = parsed.colors?.[index]?.hex;
-    return typeof hex === 'string' ? normalizeDesignKitHex(hex) : null;
-  } catch {
-    return null;
-  }
-}
-
-function colorHexFromDesignMd(body: string, index: number): string | null {
-  if (!body.trim()) return null;
-  return normalizeDesignKitHex(parseDesignMd(body).colors[index]?.hex ?? '');
-}
-
-function designMdBodyWithColor(
-  body: string,
-  colors: KitColor[],
-  index: number,
-  hex: string,
-): string {
-  const replaced = replaceDesignMdColorAtIndex(body, index, hex);
-  if (replaced) return replaced;
-  const nextColors = colors.length > 0
-    ? colors.map((color, colorIndex) => ({
-        ...color,
-        hex: colorIndex === index ? hex : color.hex,
-      }))
-    : [];
-  while (nextColors.length <= index) {
-    nextColors.push({
-      role: `color-${nextColors.length + 1}`,
-      name: `Color ${nextColors.length + 1}`,
-      hex: nextColors.length === index ? hex : '#000000',
-      usage: '',
-    });
-  }
-  if (nextColors[index]) {
-    nextColors[index] = { ...nextColors[index], hex };
-  }
-  const table = [
-    '## Color Palette',
-    '',
-    '| Role | Name | Hex | Usage |',
-    '| --- | --- | --- | --- |',
-    ...nextColors.map((color, colorIndex) => {
-      const role = color.role || `color-${colorIndex + 1}`;
-      const name = color.name || role;
-      return `| ${role} | ${name} | \`${normalizeDesignKitHex(color.hex) ?? '#000000'}\` | ${color.usage || ''} |`;
-    }),
-  ].join('\n');
-  return `${body.trimEnd()}\n\n${table}\n`;
-}
-
-function designSystemHasSourceContext(system: DesignSystemSummary): boolean {
-  const provenance = system.provenance;
-  if (!provenance) return false;
-  return Boolean(
-    provenance.companyBlurb?.trim() ||
-    provenance.githubUrls?.length ||
-    provenance.localCodeFiles?.length ||
-    provenance.figFiles?.length ||
-    provenance.assetFiles?.length ||
-    provenance.notes?.trim() ||
-    provenance.sourceNotes?.trim(),
-  );
-}
-
-function slugForTestId(value: string): string {
-  return value.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/^-+|-+$/g, '');
-}
-
-function designSystemSectionEditableFile(
-  section: DesignSystemProjectSection,
-  previewFile: ProjectFile | null,
-  fileByName: Map<string, ProjectFile>,
-): ProjectFile | null {
-  if (previewFile && (previewFile.kind === 'html' || previewFile.kind === 'sketch')) return previewFile;
-  const htmlFile = section.files
-    .map((name) => fileByName.get(name))
-    .find((file) => file?.kind === 'html');
-  if (htmlFile) return htmlFile;
-  return previewFile ?? section.files.map((name) => fileByName.get(name)).find(Boolean) ?? null;
-}
-
-function designSystemSectionPreviewFile(
-  names: string[],
-  fileByName: Map<string, ProjectFile>,
-): ProjectFile | null {
-  for (const name of names) {
-    const file = fileByName.get(name);
-    if (!file) continue;
-    if (file.kind === 'html' || file.kind === 'image' || file.kind === 'sketch') return file;
-  }
-  return null;
-}
-
-function buildDesignSystemReviewSections(
-  names: string[],
-  fileByName: Map<string, ProjectFile>,
-  cardManifest: DesignSystemCardManifestMap = new Map(),
-): DesignSystemProjectSection[] {
-  const artifactNames = names
-    .filter((name) => isDesignSystemReviewArtifactFile(name, fileByName))
-    .sort(designSystemReviewArtifactSort);
-  if (artifactNames.length > 0) {
-    const reviewNames = preferPreviewArtifactsOverRawAssets(artifactNames);
-    return reviewNames.map((name) => {
-      const manifestEntry = cardManifest.get(normalizeDesignSystemPath(name));
-      const title = manifestEntry?.name?.trim() || designSystemReviewTitleFromPath(name);
-      const category = inferDesignSystemReviewCategory(name, title, manifestEntry);
-      return {
-        title,
-        subtitle: manifestEntry?.subtitle?.trim() || designSystemReviewSubtitle(title, category, name),
-        category,
-        files: designSystemRelatedFilesForCategory(name, category, names),
-      };
-    });
-  }
-  return designSystemFallbackReviewSections(names);
-}
-
-function preferPreviewArtifactsOverRawAssets(names: string[]): string[] {
-  const hasBrandPreview = names.some((name) => {
-    const path = normalizeDesignSystemPath(name);
-    const title = designSystemReviewTitleFromPath(name);
-    return inferDesignSystemReviewCategory(name, title) === 'Brand'
-      && (path.startsWith('preview/') || path.includes('/preview/') || path.endsWith('.html'));
-  });
-  if (!hasBrandPreview) return names;
-  return names.filter((name) => {
-    const path = normalizeDesignSystemPath(name);
-    const title = designSystemReviewTitleFromPath(name);
-    if (inferDesignSystemReviewCategory(name, title) !== 'Brand') return true;
-    return path.startsWith('preview/') || path.includes('/preview/') || path.endsWith('.html');
-  });
-}
-
-function isDesignSystemReviewArtifactFile(
-  name: string,
-  fileByName: Map<string, ProjectFile>,
-): boolean {
-  const path = normalizeDesignSystemPath(name);
-  const file = fileByName.get(name);
-  if (!file || isDesignSystemEvidenceFile(path) || path === 'metadata.json') return false;
-  const isRenderable = file.kind === 'html' || file.kind === 'image' || file.kind === 'sketch';
-  if (!isRenderable) return false;
-  if (isDesignSystemRawAssetFile(path)) return isDesignSystemReviewableAssetArtifact(path);
-  if (path === 'index.html') return true;
-  if (path.startsWith('preview/') || path.includes('/preview/')) return true;
-  if (isDesignSystemUiKitFile(path)) return true;
-  return false;
-}
-
-function isDesignSystemRawAssetFile(path: string): boolean {
-  return path.startsWith('assets/')
-    || path.startsWith('src/assets/')
-    || path.startsWith('public/')
-    || path.includes('/assets/')
-    || path.includes('/src/assets/')
-    || path.includes('/fonts/')
-    || path.includes('/logos/');
-}
-
-function isDesignSystemReviewableAssetArtifact(path: string): boolean {
-  return /\b(brand|logo|logos|mark|wordmark|icon)\b/u.test(path);
-}
-
-function formatWorkspaceSnapshotElapsed(seconds: number): string {
-  const safe = Math.max(0, Math.floor(seconds));
-  if (safe < 60) return `${safe}s`;
-  const minutes = Math.floor(safe / 60);
-  const remainder = safe % 60;
-  return remainder === 0 ? `${minutes}m` : `${minutes}m ${String(remainder).padStart(2, '0')}s`;
-}
-
-function designSystemReviewArtifactSort(first: string, second: string): number {
-  const firstCategory = inferDesignSystemReviewCategory(first, designSystemReviewTitleFromPath(first));
-  const secondCategory = inferDesignSystemReviewCategory(second, designSystemReviewTitleFromPath(second));
-  return designSystemReviewCategoryRank(firstCategory) - designSystemReviewCategoryRank(secondCategory)
-    || designSystemReviewTitleFromPath(first).localeCompare(designSystemReviewTitleFromPath(second));
-}
-
-function designSystemReviewTitleFromPath(name: string): string {
-  const path = normalizeDesignSystemPath(name);
-  const parts = path.split('/').filter(Boolean);
-  let basename = parts[parts.length - 1] ?? path;
-  if (/^index\.(html?|png|jpe?g|svg|webp|avif)$/iu.test(basename) && parts.length > 1) {
-    basename = parts[parts.length - 2] ?? basename;
-  }
-  return basename
-    .replace(/\.(html?|png|jpe?g|gif|webp|avif|svg|fig|pen)$/iu, '')
-    .replace(/_/g, '-')
-    .replace(/\s+/g, '-')
-    .replace(/^-+|-+$/g, '')
-    || 'overview';
-}
-
-function inferDesignSystemReviewCategory(
-  name: string,
-  title: string,
-  manifestEntry?: DesignSystemCardManifestEntry,
-): DesignSystemReviewCategory {
-  const text = `${normalizeDesignSystemPath(name)} ${title}`.toLowerCase();
-  const group = manifestEntry?.group?.toLowerCase() ?? '';
-  if (group.includes('ui kit')) return 'Components';
-  if (/\b(type|typography|font|text)\b/u.test(text)) return 'Type';
-  if (/\b(color|colors|palette|theme)\b/u.test(text)) return 'Colors';
-  if (/\b(space|spacing|radius|radii|shadow|shadows|elevation|layout-grid)\b/u.test(text)) return 'Spacing';
-  if (/\b(brand|logo|logos|mark|wordmark|icon|favicon)\b/u.test(text)) return 'Brand';
-  if (group.includes('brand')) return 'Brand';
-  return 'Components';
-}
-
-function designSystemReviewSubtitle(title: string, category: DesignSystemReviewCategory, name = ''): string {
-  const path = normalizeDesignSystemPath(name);
-  const titleText = title.toLowerCase();
-  const text = `${title} ${path}`.toLowerCase();
-  if (isDesignSystemUiKitEntryPage(path)) return 'Applied UI kit example';
-  if (text.includes('typography')) return 'Text hierarchy and styles';
-  if (text.includes('type-')) return 'Typography scale and font guidance';
-  if (text.includes('font')) return 'Font family specimens';
-  if (text.includes('node')) return 'Data type color coding system';
-  if (text.includes('ui-palette') || text.includes('palette')) return 'Interface color palette';
-  if (text.includes('dark')) return 'Dark theme color palette';
-  if (text.includes('spacing') || text.includes('radius') || text.includes('radii') || text.includes('shadow')) return 'Spacing scale and border radius tokens';
-  if (text.includes('favicon')) return 'Brand app icon and favicon';
-  if (text.includes('logo') || text.includes('brand')) return 'Brand logo marks';
-  if (titleText.includes('interface') || titleText.includes('ui')) return 'Interface and component patterns';
-  switch (category) {
-    case 'Type':
-      return 'Typography scale and font guidance';
-    case 'Colors':
-      return 'Color palette and token specimens';
-    case 'Spacing':
-      return 'Spacing and radius system';
-    case 'Brand':
-      return 'Brand assets and identity usage';
-    case 'Components':
-      return 'Reusable product interface examples';
-  }
-}
-
-function isDesignSystemUiKitEntryPage(path: string): boolean {
-  return isDesignSystemUiKitFile(path) && /\.html?$/iu.test(path);
-}
-
-function designSystemManifestCardError(index: number, detail: string): Error {
-  const separator = detail.startsWith('.') ? '' : ' ';
-  return new Error(`Invalid _ds_manifest.json: cards[${index}]${separator}${detail}.`);
-}
-
-function optionalDesignSystemManifestString(
-  record: Record<string, unknown>,
-  field: (typeof DESIGN_SYSTEM_CARD_MANIFEST_OPTIONAL_STRING_FIELDS)[number],
-  index: number,
-): string | undefined {
-  const value = record[field];
-  if (value === undefined) return undefined;
-  if (typeof value !== 'string') throw designSystemManifestCardError(index, `.${field} must be a string`);
-  return value;
-}
-
-function parseDesignSystemCardManifestEntry(card: unknown, index: number): DesignSystemCardManifestEntry {
-  if (!card || typeof card !== 'object' || Array.isArray(card)) {
-    throw designSystemManifestCardError(index, 'must be an object');
-  }
-  const record = card as Record<string, unknown>;
-  if (typeof record.path !== 'string' || !record.path.trim()) {
-    throw designSystemManifestCardError(index, '.path must be a non-empty string');
-  }
-  const entry: DesignSystemCardManifestEntry = { path: normalizeDesignSystemPath(record.path) };
-  for (const field of DESIGN_SYSTEM_CARD_MANIFEST_OPTIONAL_STRING_FIELDS) {
-    entry[field] = optionalDesignSystemManifestString(record, field, index);
-  }
-  return entry;
-}
-
-function parseDesignSystemCardManifest(text: string | null): DesignSystemCardManifestMap {
-  if (!text) return new Map();
-  let parsed: { cards?: unknown };
-  try {
-    parsed = JSON.parse(text) as { cards?: unknown };
-  } catch (err: unknown) {
-    const detail = err instanceof Error ? err.message : String(err);
-    throw new Error(`Invalid _ds_manifest.json: ${detail}`);
-  }
-  if (!parsed || typeof parsed !== 'object') {
-    throw new Error('Invalid _ds_manifest.json: expected an object with a cards array.');
-  }
-  if (parsed.cards !== undefined && !Array.isArray(parsed.cards)) {
-    throw new Error('Invalid _ds_manifest.json: cards must be an array.');
-  }
-  const cards = Array.isArray(parsed.cards) ? parsed.cards : [];
-  const entries: Array<[string, DesignSystemCardManifestEntry]> = [];
-  for (const [index, card] of cards.entries()) {
-    const entry = parseDesignSystemCardManifestEntry(card, index);
-    entries.push([entry.path, entry]);
-  }
-  return new Map(entries);
-}
-
-function designSystemReviewPreviewDisplay(
-  section: DesignSystemProjectSection,
-  previewFile: ProjectFile | null,
-): DesignSystemReviewPreviewDisplay {
-  if (!previewFile) return 'specimen';
-  const path = normalizeDesignSystemPath(previewFile.name);
-  if (path.startsWith('ui_kits/') || path.includes('/ui_kits/')) return 'ui-kit';
-  if (previewFile.kind !== 'html') return 'asset';
-  if (section.category === 'Components' && !path.startsWith('preview/')) return 'ui-kit';
-  return 'specimen';
-}
-
-function designSystemRelatedFilesForCategory(
-  artifactName: string,
-  category: DesignSystemReviewCategory,
-  names: string[],
-): string[] {
-  const related = names.filter((name) => {
-    if (name === artifactName || isDesignSystemEvidenceFile(name)) return false;
-    switch (category) {
-      case 'Type':
-      case 'Colors':
-      case 'Spacing':
-        return isDesignSystemTokenFile(name);
-      case 'Components':
-        return isDesignSystemUiKitFile(name);
-      case 'Brand':
-        return isDesignSystemAssetFile(name);
-    }
-  });
-  return Array.from(new Set([artifactName, ...related])).slice(0, 12);
-}
-
-function designSystemFallbackReviewSections(names: string[]): DesignSystemProjectSection[] {
-  const tokenFiles = names.filter(isDesignSystemTokenFile).slice(0, 8);
-  const uiKitFiles = names.filter(isDesignSystemUiKitFile).slice(0, 8);
-  const assetFiles = names.filter(isDesignSystemAssetFile).slice(0, 8);
-  const sections: Array<DesignSystemProjectSection | null> = [
-    tokenFiles.length > 0
-      ? {
-        title: 'colors-and-type',
-        subtitle: 'Color, type, spacing, and token guidance',
-        category: 'Colors',
-        files: tokenFiles,
-      }
-      : null,
-    uiKitFiles.length > 0
-      ? {
-        title: 'components',
-        subtitle: 'Reusable interface examples',
-        category: 'Components',
-        files: uiKitFiles,
-      }
-      : null,
-    assetFiles.length > 0
-      ? {
-        title: 'assets',
-        subtitle: 'Brand logos, fonts, and uploaded assets',
-        category: 'Brand',
-        files: assetFiles,
-      }
-      : null,
-  ];
-  return sections.filter((section): section is DesignSystemProjectSection => section !== null);
-}
-
-function designSystemReviewGroups(
-  reviews: DesignSystemProjectSectionReview[],
-): Array<{ title: DesignSystemReviewCategory; items: DesignSystemProjectSectionReview[] }> {
-  const categories: DesignSystemReviewCategory[] = ['Type', 'Colors', 'Spacing', 'Components', 'Brand'];
-  return categories
-    .map((title) => ({
-      title,
-      items: reviews.filter((review) => review.section.category === title),
-    }))
-    .filter((group) => group.items.length > 0);
-}
-
-function designSystemReviewCategoryRank(category: DesignSystemReviewCategory): number {
-  return ['Type', 'Colors', 'Spacing', 'Components', 'Brand'].indexOf(category);
-}
-
-function designSystemReviewNeedsAttention(review: DesignSystemProjectSectionReview): boolean {
-  return review.sectionStatus === 'needs-review'
-    || review.sectionStatus === 'needs-work'
-    || review.sectionStatus === 'updated'
-    || review.sectionStatus === 'running'
-    || review.sectionStatus === 'planned'
-    || review.sectionStatus === 'missing';
-}
-
-function isDesignSystemEvidenceFile(name: string): boolean {
-  const path = normalizeDesignSystemPath(name);
-  return path.startsWith('context/') || path.includes('/context/');
-}
-
-function isDesignSystemGuidanceFile(name: string): boolean {
-  const path = normalizeDesignSystemPath(name);
-  if (path.includes('/')) return false;
-  return DESIGN_SYSTEM_GUIDANCE_FILES.has(path);
-}
-
-function designSystemGuidanceSort(first: string, second: string): number {
-  const order = ['design.md', 'readme.md', 'readme-print.md', 'skill.md'];
-  const firstRank = order.indexOf(normalizeDesignSystemPath(first));
-  const secondRank = order.indexOf(normalizeDesignSystemPath(second));
-  return (firstRank === -1 ? order.length : firstRank)
-    - (secondRank === -1 ? order.length : secondRank)
-    || first.localeCompare(second);
-}
-
-function isDesignSystemTokenFile(name: string): boolean {
-  const path = normalizeDesignSystemPath(name);
-  if (isDesignSystemEvidenceFile(path)) return false;
-  if (
-    path.startsWith('preview/')
-    || path.startsWith('ui_kits/')
-    || path.startsWith('assets/')
-    || path.startsWith('src/assets/')
-    || path.startsWith('public/')
-    || path.includes('/preview/')
-    || path.includes('/ui_kits/')
-    || path.includes('/assets/')
-    || path.includes('/src/assets/')
-    || DESIGN_SYSTEM_IMAGE_OR_FONT_EXTENSIONS.test(path)
-  ) {
-    return false;
-  }
-  const basename = designSystemBasename(path);
-  if (basename.endsWith('.html')) return false;
-  return basename === 'colors_and_type.css'
-    || basename === 'tailwind.config.ts'
-    || basename === 'tailwind.config.js'
-    || basename === 'tailwind.config.mjs'
-    || basename === 'theme.css'
-    || basename === 'tokens.css'
-    || basename === 'variables.css'
-    || basename === 'design-tokens.json'
-    || path.includes('/tokens/')
-    || path.startsWith('src/tokens/')
-    || path.startsWith('src/styles/')
-    || path.startsWith('styles/')
-    || /\b(color|colors|palette|typography|spacing|radius|theme|token)s?\b/u.test(path);
-}
-
-function isDesignSystemPreviewFile(name: string): boolean {
-  const path = normalizeDesignSystemPath(name);
-  if (isDesignSystemEvidenceFile(path) || path.startsWith('ui_kits/')) return false;
-  const basename = designSystemBasename(path);
-  return path.startsWith('preview/')
-    || (path.split('/').length === 1 && basename.endsWith('.html'))
-    || (basename.endsWith('.html') && /\b(index|overview|preview|showcase|styleguide)\b/u.test(path));
-}
-
-function isDesignSystemUiKitFile(name: string): boolean {
-  const path = normalizeDesignSystemPath(name);
-  if (isDesignSystemEvidenceFile(path)) return false;
-  if (isDesignSystemRawAssetFile(path)) return false;
-  return path.startsWith('ui_kits/')
-    || path.startsWith('src/components/')
-    || path.startsWith('components/')
-    || path.includes('/ui_kits/')
-    || path.includes('/src/components/')
-    || /\b(component|components|interface|ui-kit|uikit)\b/u.test(path);
-}
-
-function isDesignSystemAssetFile(name: string): boolean {
-  const path = normalizeDesignSystemPath(name);
-  if (isDesignSystemEvidenceFile(path)) return false;
-  return path.startsWith('assets/')
-    || path.startsWith('src/assets/')
-    || path.startsWith('public/')
-    || path.includes('/assets/')
-    || path.includes('/src/assets/')
-    || path.includes('/fonts/')
-    || path.includes('/icons/')
-    || path.includes('/logos/')
-    || DESIGN_SYSTEM_IMAGE_OR_FONT_EXTENSIONS.test(path);
-}
-
-function designSystemGenerationReviewHasStarted(
-  sectionReviews: DesignSystemProjectSectionReview[],
-): boolean {
-  return sectionReviews.some((review) => {
-    const { previewFile, section, sectionActivity } = review;
-    if (previewFile) return true;
-    if (section.files.length > 0 && sectionActivity.phase !== 'idle') return true;
-    return sectionActivity.phase === 'writing'
-      || sectionActivity.phase === 'updated'
-      || sectionActivity.phase === 'planned';
-  });
-}
-
-function designSystemSectionVisibleDuringGeneration(
-  review: DesignSystemProjectSectionReview,
-): boolean {
-  const { section, reviewEntry, sectionActivity, previewFile } = review;
-  if (reviewEntry) return true;
-  if (previewFile) return true;
-  if (sectionActivity.phase !== 'idle') return true;
-  return section.files.length > 0;
-}
-
-function designSystemSectionStatus(
-  section: DesignSystemProjectSection,
-  decision: DesignSystemReviewDecision | undefined,
-  changedAfterFeedback: boolean,
-  activity: DesignSystemSectionActivity,
-): DesignSystemSectionStatus {
-  if (activity.running) return 'running';
-  if (activity.phase === 'planned') return 'planned';
-  if (changedAfterFeedback || activity.mutated) return 'updated';
-  if (section.files.length === 0) return 'missing';
-  if (decision === 'looks-good') return 'approved';
-  if (decision === 'needs-work') return 'needs-work';
-  return 'needs-review';
-}
-
-function designSystemSectionStatusLabel(
-  t: TranslateFn,
-  section: DesignSystemProjectSection,
-  status: DesignSystemSectionStatus,
-  activity: DesignSystemSectionActivity,
-): string {
-  switch (status) {
-    case 'running':
-      return designSystemSectionPhaseLabel(t, section, activity);
-    case 'planned':
-      return t('ds.sectionQueued');
-    case 'updated':
-      return t('ds.sectionReviewUpdatedFiles');
-    case 'approved':
-      return t('ds.reviewLooksGood');
-    case 'needs-work':
-      return t('ds.reviewNeedsWork');
-    case 'needs-review':
-      return t('ds.reviewNeedsReview');
-    case 'missing':
-      return section.requiredFile
-        ? t('ds.sectionRequiredFileMissing', { file: section.requiredFile })
-        : t('ds.sectionNoFilesYet');
-  }
-}
-
-function designSystemSectionStatusClass(status: DesignSystemSectionStatus): string {
-  switch (status) {
-    case 'running':
-      return 'is-running';
-    case 'planned':
-      return 'is-planned';
-    case 'updated':
-      return 'is-review';
-    case 'approved':
-      return 'is-approved';
-    case 'needs-work':
-      return 'is-work';
-    case 'needs-review':
-      return 'is-ready';
-    case 'missing':
-      return 'is-missing';
-  }
-}
-
-function designSystemInitialGenerationSteps({
-  files,
-  sectionReviews,
-  system,
-  t,
-}: {
-  files: ProjectFile[];
-  sectionReviews: DesignSystemProjectSectionReview[];
-  system: DesignSystemSummary;
-  t: TranslateFn;
-}): DesignSystemGenerationStep[] {
-  const hasSourceContext =
-    designSystemGithubEvidenceState(system, files.map((file) => file.name)).ready
-    && (
-      files.some((file) => normalizeDesignSystemPath(file.name).startsWith('context/')) ||
-      designSystemHasSourceContext(system)
-    );
-  const fileNames = files.map((file) => file.name);
-  const categoryHasReview = (category: DesignSystemReviewCategory) =>
-    sectionReviews.some((review) => review.section.category === category);
-  const categoryIsRunning = (category: DesignSystemReviewCategory) =>
-    sectionReviews.some((review) => review.section.category === category && review.sectionActivity.running);
-  const guidanceRunning = sectionReviews.some((review) =>
-    review.sectionActivity.running
-    && review.section.files.some((name) => isDesignSystemGuidanceFile(name)),
-  );
-  const steps: DesignSystemGenerationStep[] = [
-    {
-      id: 'source-context',
-      title: t('ds.generationSourceTitle'),
-      detail: t('ds.generationSourceDetail'),
-      status: hasSourceContext ? 'succeeded' : 'running',
-    },
-    {
-      id: 'guidance',
-      title: t('ds.generationGuidanceTitle'),
-      detail: t('ds.generationGuidanceDetail'),
-      status: fileNames.some(isDesignSystemGuidanceFile)
-        ? 'succeeded'
-        : guidanceRunning
-          ? 'running'
-          : 'pending',
-    },
-    {
-      id: 'tokens',
-      title: t('ds.generationTokensTitle'),
-      detail: t('ds.generationTokensDetail'),
-      status: fileNames.some(isDesignSystemTokenFile)
-        ? 'succeeded'
-        : (categoryIsRunning('Type') || categoryIsRunning('Colors') || categoryIsRunning('Spacing'))
-          ? 'running'
-          : 'pending',
-    },
-    {
-      id: 'previews',
-      title: t('ds.generationPreviewsTitle'),
-      detail: t('ds.generationPreviewsDetail'),
-      status: sectionReviews.some((review) => review.previewFile)
-        ? 'succeeded'
-        : (categoryIsRunning('Type') || categoryIsRunning('Colors') || categoryIsRunning('Spacing') || categoryIsRunning('Brand'))
-          ? 'running'
-          : 'pending',
-    },
-    {
-      id: 'ui-kit',
-      title: t('ds.generationUiKitTitle'),
-      detail: t('ds.generationUiKitDetail'),
-      status: categoryHasReview('Components') || fileNames.some(isDesignSystemUiKitFile)
-        ? 'succeeded'
-        : categoryIsRunning('Components')
-          ? 'running'
-          : 'pending',
-    },
-    {
-      id: 'assets',
-      title: t('ds.generationAssetsTitle'),
-      detail: t('ds.generationAssetsDetail'),
-      status: categoryHasReview('Brand') || fileNames.some(isDesignSystemAssetFile)
-        ? 'succeeded'
-        : categoryIsRunning('Brand')
-          ? 'running'
-          : 'pending',
-    },
-  ];
-  if (!steps.some((step) => step.status === 'running')) {
-    const firstPending = steps.find((step) => step.status === 'pending');
-    if (firstPending) firstPending.status = 'running';
-  }
-  return steps;
-}
-
-function designSystemGenerationProgress(steps: DesignSystemGenerationStep[]): number {
-  if (steps.length === 0) return 8;
-  const succeeded = steps.filter((step) => step.status === 'succeeded').length;
-  const running = steps.some((step) => step.status === 'running') ? 0.45 : 0;
-  return Math.max(8, Math.min(92, Math.round(((succeeded + running) / steps.length) * 100)));
-}
-
-function designSystemSectionActivity(
-  section: DesignSystemProjectSection,
-  fileOps: FileOpEntry[],
-  todos: TodoItem[],
-): DesignSystemSectionActivity {
-  const touched = fileOps.filter((entry) => designSystemFileOpBelongsToSection(entry, section));
-  const touchedFiles = Array.from(new Set(touched.map((entry) => entry.path)));
-  const todo = designSystemSectionTodo(section, todos);
-  const hasRunningMutation = touched.some((entry) =>
-    entry.status === 'running' && (entry.ops.includes('write') || entry.ops.includes('edit')),
-  );
-  const hasRunningRead = touched.some((entry) =>
-    entry.status === 'running' && entry.ops.includes('read'),
-  );
-  const mutated = touched.some((entry) =>
-    entry.status === 'done' && (entry.ops.includes('write') || entry.ops.includes('edit')),
-  );
-  const errored = touched.some((entry) => entry.status === 'error');
-  const todoPhase = todo ? designSystemTodoActivityPhase(section, todo) : null;
-  const hasRunningTodo = todo?.status === 'in_progress';
-  const phase: DesignSystemSectionActivityPhase =
-    errored
-      ? 'error'
-      : hasRunningMutation
-        ? 'writing'
-        : hasRunningRead
-          ? 'reading'
-          : hasRunningTodo && todoPhase
-            ? todoPhase
-            : mutated
-              ? 'updated'
-              : todoPhase
-                ? todoPhase
-                : 'idle';
-  return {
-    running: hasRunningMutation || hasRunningRead || hasRunningTodo,
-    mutated,
-    errored,
-    phase,
-    touchedFiles,
-    todoText: todo?.content,
-    todoStatus: todo?.status,
-  };
-}
-
-function designSystemSectionTodo(
-  section: DesignSystemProjectSection,
-  todos: TodoItem[],
-): TodoItem | undefined {
-  return todos
-    .filter((todo) => todo.status !== 'completed')
-    .filter((todo) => designSystemTodoBelongsToSection(todo, section))
-    .sort((first, second) => designSystemTodoRank(first) - designSystemTodoRank(second))[0];
-}
-
-function designSystemTodoRank(todo: TodoItem): number {
-  if (todo.status === 'in_progress') return 0;
-  if (todo.status === 'pending') return 1;
-  return 2;
-}
-
-function designSystemTodoActivityPhase(
-  section: DesignSystemProjectSection,
-  todo: TodoItem,
-): DesignSystemSectionActivityPhase {
-  if (todo.status === 'pending') return 'planned';
-  const text = designSystemTodoSearchText(todo);
-  const isMutation = [
-    'build',
-    'copy',
-    'create',
-    'edit',
-    'generate',
-    'import',
-    'register',
-    'update',
-    'write',
-  ].some((keyword) => text.includes(keyword));
-  if (isMutation) return 'writing';
-  const isReading = [
-    'analy',
-    'browse',
-    'explore',
-    'fetch',
-    'github',
-    'inspect',
-    'read',
-    'repo',
-    'search',
-  ].some((keyword) => text.includes(keyword));
-  if (isReading) return 'reading';
-  return section.title === 'Preview' || section.title === 'UI kit' ? 'writing' : 'reading';
-}
-
-function designSystemTodoBelongsToSection(
-  todo: TodoItem,
-  section: DesignSystemProjectSection,
-): boolean {
-  const text = designSystemTodoSearchText(todo);
-  if (section.files.some((name) => text.includes(designSystemReviewTitleFromPath(name)))) {
-    return true;
-  }
-  switch (section.category) {
-    case 'Type':
-      return [
-        'font',
-        'type',
-        'typography',
-      ].some((keyword) => text.includes(keyword));
-    case 'Colors':
-      return [
-        'color',
-        'colors_and_type',
-        'css variable',
-        'palette',
-        'theme',
-        'token',
-      ].some((keyword) => text.includes(keyword));
-    case 'Spacing':
-      return [
-        'radius',
-        'spacing',
-        'space',
-      ].some((keyword) => text.includes(keyword));
-    case 'Components':
-      return [
-        'component',
-        'interface',
-        'prototype',
-        'react',
-        'ui kit',
-        'ui_kit',
-        'ui_kits',
-      ].some((keyword) => text.includes(keyword));
-    case 'Brand':
-      return [
-        'font',
-        'icon',
-        'logo',
-        'brand',
-        'asset',
-        'upload',
-      ].some((keyword) => text.includes(keyword));
-  }
-}
-
-function designSystemTodoSearchText(todo: TodoItem): string {
-  return `${todo.content} ${todo.activeForm ?? ''}`.toLowerCase();
-}
-
-function designSystemFileOpBelongsToSection(
-  entry: FileOpEntry,
-  section: DesignSystemProjectSection,
-): boolean {
-  const candidates = [entry.fullPath, entry.path].map(normalizeDesignSystemPath);
-  const sectionFiles = [...section.files, section.requiredFile]
-    .filter((name): name is string => Boolean(name))
-    .map(normalizeDesignSystemPath);
-  if (sectionFiles.some((name) => candidates.some((candidate) =>
-    candidate === name || candidate.endsWith(`/${name}`),
-  ))) {
-    return true;
-  }
-  return candidates.some((path) => designSystemPathMatchesSection(path, section.category));
-}
-
-function designSystemPathMatchesSection(path: string, sectionTitle: string): boolean {
-  const basename = designSystemBasename(path);
-  switch (sectionTitle) {
-    case 'Type':
-      return !isDesignSystemEvidenceFile(path)
-        && (isDesignSystemTokenFile(path) || DESIGN_SYSTEM_GUIDANCE_FILES.has(basename))
-        && /\b(type|typography|font|text)\b/u.test(path);
-    case 'Colors':
-      return isDesignSystemTokenFile(path)
-        && /\b(color|colors|palette|theme|token)\b/u.test(path);
-    case 'Spacing':
-      return isDesignSystemTokenFile(path)
-        && /\b(space|spacing|radius)\b/u.test(path);
-    case 'Components':
-      return isDesignSystemUiKitFile(path);
-    case 'Brand':
-      return isDesignSystemAssetFile(path);
-    default:
-      return false;
-  }
-}
-
-function normalizeDesignSystemPath(path: string): string {
-  return path.replace(/\\/g, '/').replace(/^\.?\//, '').toLowerCase();
-}
-
-function normalizeProjectFilePath(path: string): string {
-  return path.replace(/\\/g, '/').split('/').filter(Boolean).join('/');
-}
-
-function joinProjectFilePath(dir: string, name: string): string {
-  const normalizedDir = normalizeProjectFilePath(dir);
-  return normalizedDir ? `${normalizedDir}/${name}` : name;
-}
-
-function nextMarkdownDocumentPath(files: ProjectFile[], dir: string): string {
-  const existing = new Set(files.map((file) => normalizeProjectFilePath(file.name).toLowerCase()));
-  for (let index = 1; index < 1000; index += 1) {
-    const name = index === 1 ? 'document.md' : `document-${index}.md`;
-    const candidate = joinProjectFilePath(dir, name);
-    if (!existing.has(normalizeProjectFilePath(candidate).toLowerCase())) return candidate;
-  }
-  return joinProjectFilePath(dir, `document-${Date.now()}.md`);
-}
-
-function initialMarkdownDocument(
-  path: string,
-  projectKind: TrackingProjectKind,
-  t: TranslateFn,
-): string {
-  const title = normalizeProjectFilePath(path)
-    .split('/')
-    .pop()
-    ?.replace(/\.mdx?$/i, '')
-    .replace(/[-_]+/g, ' ')
-    .replace(/\b\w/g, (char) => char.toUpperCase()) || t('designFiles.documentTemplate.titleFallback');
-  return `# ${title}
-
-## ${t('designFiles.documentTemplate.goalHeading')}
-
-${t('designFiles.documentTemplate.goalBody')}
-
-## ${t('designFiles.documentTemplate.capabilitiesHeading')}
-
-- ${t('designFiles.documentTemplate.capabilityMarkdown')}
-- ${t('designFiles.documentTemplate.capabilityAgent')}
-- ${t('designFiles.documentTemplate.capabilityImages')}
-
-## ${t('designFiles.documentTemplate.scenarioHeading')}
-
-${t(documentTemplateScenarioKey(projectKind))}
-
-## ${t('designFiles.documentTemplate.nextHeading')}
-
-${t('designFiles.documentTemplate.nextBody')}
-`;
-}
-
-function documentTemplateScenarioKey(projectKind: TrackingProjectKind): keyof Dict {
-  switch (projectKind) {
-    case 'prototype':
-      return 'designFiles.documentTemplate.scenario.prototype';
-    case 'wireframe':
-      return 'designFiles.documentTemplate.scenario.wireframe';
-    case 'mobile':
-      return 'designFiles.documentTemplate.scenario.mobile';
-    case 'slide_deck':
-      return 'designFiles.documentTemplate.scenario.slideDeck';
-    case 'document':
-      return 'designFiles.documentTemplate.scenario.document';
-    case 'image':
-      return 'designFiles.documentTemplate.scenario.image';
-    case 'video':
-      return 'designFiles.documentTemplate.scenario.video';
-    case 'hyperframes':
-      return 'designFiles.documentTemplate.scenario.hyperframes';
-    case 'audio':
-      return 'designFiles.documentTemplate.scenario.audio';
-    case 'live_artifact':
-      return 'designFiles.documentTemplate.scenario.liveArtifact';
-    case 'brand':
-    case 'design_system':
-      return 'designFiles.documentTemplate.scenario.designSystem';
-    case 'template':
-    default:
-      return 'designFiles.documentTemplate.scenario.default';
-  }
-}
-
-function designSystemBasename(path: string): string {
-  const segments = normalizeDesignSystemPath(path).split('/').filter(Boolean);
-  return segments[segments.length - 1] ?? normalizeDesignSystemPath(path);
-}
-
-function designSystemSectionPhaseLabel(
-  t: TranslateFn,
-  section: DesignSystemProjectSection,
-  activity: DesignSystemSectionActivity,
-): string {
-  if (activity.phase === 'planned') {
-    switch (section.category) {
-      case 'Type':
-        return t('ds.phaseQueuedTypography');
-      case 'Colors':
-        return t('ds.phaseQueuedTokens');
-      case 'Spacing':
-        return t('ds.phaseQueuedSpacing');
-      case 'Components':
-        return t('ds.phaseQueuedUiKit');
-      case 'Brand':
-        return t('ds.phaseQueuedAssets');
-    }
-  }
-  if (activity.phase === 'reading') {
-    switch (section.category) {
-      case 'Type':
-        return t('ds.phaseReadingTypography');
-      case 'Colors':
-        return t('ds.phaseReadingTokens');
-      case 'Spacing':
-        return t('ds.phaseReadingSpacing');
-      case 'Components':
-        return t('ds.phaseReadingUiKit');
-      case 'Brand':
-        return t('ds.phaseReadingAssets');
-    }
-  }
-  if (activity.phase === 'writing') {
-    switch (section.category) {
-      case 'Type':
-        return t('ds.phaseWritingTypography');
-      case 'Colors':
-        return t('ds.phaseWritingTokens');
-      case 'Spacing':
-        return t('ds.phaseWritingSpacing');
-      case 'Components':
-        return t('ds.phaseBuildingUiKit');
-      case 'Brand':
-        return t('ds.phaseUpdatingAssets');
-    }
-  }
-  if (activity.phase === 'error') return t('ds.phaseNeedsAttention');
-  if (activity.phase === 'updated') return t('ds.phaseUpdated');
-  return t('ds.reviewNeedsReview');
-}
-
-function designSystemSectionActivityLabel(
-  t: TranslateFn,
-  section: DesignSystemProjectSection,
-  activity: DesignSystemSectionActivity,
-): string {
-  if (activity.touchedFiles.length === 0) {
-    const phaseLabel = designSystemSectionPhaseLabel(t, section, activity);
-    return activity.todoText
-      ? t('ds.sectionActivityFromTodo', {
-          phase: phaseLabel,
-          todo: truncateDesignSystemActivityText(activity.todoText),
-        })
-      : phaseLabel;
-  }
-  const label = activity.touchedFiles.slice(0, 3).join(', ');
-  const suffix = activity.touchedFiles.length > 3 ? ` +${activity.touchedFiles.length - 3}` : '';
-  const files = `${label}${suffix}`;
-  if (activity.phase === 'idle') return t('ds.sectionActivityReadFiles', { files });
-  return t('ds.sectionActivityPhaseFiles', {
-    phase: designSystemSectionPhaseLabel(t, section, activity),
-    files,
-  });
-}
-
-function truncateDesignSystemActivityText(value: string): string {
-  const trimmed = value.trim();
-  return trimmed.length > 80 ? `${trimmed.slice(0, 77)}...` : trimmed;
-}
-
-function designSystemSectionRunningNotice(
-  t: TranslateFn,
-  section: DesignSystemProjectSection,
-  activity: DesignSystemSectionActivity,
-): string {
-  if (activity.phase === 'reading') {
-    return t('ds.sectionRunningReadingContext', { title: section.title });
-  }
-  return t('ds.sectionRunningNow', { phase: designSystemSectionPhaseLabel(t, section, activity) });
-}
-
-function designSystemReviewTimeLabel(t: TranslateFn, value: string): string | null {
-  const time = Date.parse(value);
-  if (!Number.isFinite(time)) return null;
-  const formatted = new Intl.DateTimeFormat(undefined, {
-    month: 'short',
-    day: 'numeric',
-    hour: 'numeric',
-    minute: '2-digit',
-  }).format(new Date(time));
-  return t('ds.reviewLastReviewed', { time: formatted });
-}
-
-function designSystemReviewAgentTaskLabel(t: TranslateFn, task: DesignSystemReviewAgentTask): string {
-  switch (task.status) {
-    case 'queued':
-      return t('ds.agentFeedbackQueued');
-    case 'sent':
-      if (!task.sentAt) return t('ds.agentFeedbackSent');
-      {
-        const time = Date.parse(task.sentAt);
-        if (!Number.isFinite(time)) return t('ds.agentFeedbackSent');
-        const formatted = new Intl.DateTimeFormat(undefined, {
-          month: 'short',
-          day: 'numeric',
-          hour: 'numeric',
-          minute: '2-digit',
-        }).format(new Date(time));
-        return t('ds.agentFeedbackSentAt', { time: formatted });
-      }
-    case 'failed':
-      return task.error
-        ? t('ds.agentFeedbackFailedWithError', { error: task.error })
-        : t('ds.agentFeedbackFailed');
-  }
-  return t('ds.agentFeedbackUnknown');
-}
-
-function designSystemSectionChangedAfterReview(
-  names: string[],
-  fileByName: Map<string, ProjectFile>,
-  reviewEntry: DesignSystemReviewEntry | undefined,
-): boolean {
-  if (!reviewEntry || reviewEntry.decision !== 'needs-work') return false;
-  const reviewedAt = Date.parse(reviewEntry.updatedAt);
-  if (!Number.isFinite(reviewedAt)) return false;
-  const trackedNames: string[] = reviewEntry.files && reviewEntry.files.length > 0
-    ? reviewEntry.files
-    : names;
-  return trackedNames.some((name) => {
-    const file = fileByName.get(name);
-    return file ? file.mtime > reviewedAt : false;
-  });
 }
 
 function DesignSystemInlinePreview({
